feat(crowdnode): Align colors with other parts of UI and improve dark mode in general
