+++ DashWallet/Sources/UI/CrowdNode/New Account/NewAccountViewController.swift
@@ -80,11 +80,7 @@ final class NewAccountViewController: UIViewController, UITextViewDelegate {
 extension NewAccountViewController {
     private func configureHierarchy() {
         definesPresentationContext = true
-        view.backgroundColor = UIColor.dw_background()
-        let appearance = UINavigationBarAppearance()
-        appearance.configureWithTransparentBackground()
-        navigationItem.standardAppearance = appearance
-        navigationItem.scrollEdgeAppearance = appearance
+        view.backgroundColor = .dw_secondaryBackground()
 
         configureTermsCheckBox()
