+++ DashWallet/Sources/UI/CrowdNode/Portal/CrowdNodePortalViewController.swift
@@ -74,6 +74,8 @@ final class CrowdNodePortalController: UIViewController {
 
 extension CrowdNodePortalController {
     private func configureHierarchy() {
+        view.backgroundColor = .dw_secondaryBackground()
+
         balanceView.tint = .white
         balanceView.dataSource = self
 
@@ -96,8 +98,6 @@ extension CrowdNodePortalController {
     private func configureNavBar() {
         let appearance = UINavigationBarAppearance()
         appearance.configureWithTransparentBackground()
-        let image = UIImage(systemName: "chevron.backward")?.withTintColor(.white, renderingMode: .alwaysOriginal)
-        appearance.setBackIndicatorImage(image, transitionMaskImage: image)
         navigationItem.standardAppearance = appearance
         navigationItem.scrollEdgeAppearance = appearance
 
@@ -197,76 +197,8 @@ extension CrowdNodePortalController {
     }
 }
 
-// MARK: - CrowdNodeCell
 
-class CrowdNodeCell: UITableViewCell {
-    @IBOutlet var title : UILabel!
-    @IBOutlet var subtitle : UILabel!
-    @IBOutlet var icon : UIImageView!
-    @IBOutlet var iconCircle : UIView!
-    @IBOutlet var additionalInfo: UIView!
-    @IBOutlet var additionalInfoLabel : UILabel!
-    @IBOutlet var additionalInfoIcon : UIImageView!
-    @IBOutlet var verifyButton : UIButton!
-
-    @IBOutlet var showInfoConstraint: NSLayoutConstraint!
-    @IBOutlet var collapseInfoConstraint: NSLayoutConstraint!
-    @IBOutlet var infoBottomAnchorConstraint: NSLayoutConstraint!
-
-    fileprivate func update(with item: CrowdNodePortalItem,
-                            _ crowdNodeBalance: UInt64,
-                            _ walletBalance: UInt64,
-                            _ onlineAccountState: CrowdNode.OnlineAccountState) {
-        title.text = item.title
-        subtitle.text = item.subtitle
-        icon.image = UIImage(named: item.icon)
-
-        if item.isDisabled(crowdNodeBalance, walletBalance, onlineAccountState.isLinkingInProgress) {
-            let grayColor = UIColor(red: 176/255.0, green: 182/255.0, blue: 188/255.0, alpha: 1.0)
-            iconCircle.backgroundColor = grayColor
-            title.textColor = .dw_secondaryText()
-            selectionStyle = .none
-        } else {
-            iconCircle.backgroundColor = item.iconCircleColor
-            title.textColor = .label
-            selectionStyle = .default
-        }
-
-        var showInfo: Bool
-
-        switch item {
-        case .deposit:
-            showInfo = crowdNodeBalance < CrowdNode.minimumDeposit && !onlineAccountState.isLinkingInProgress
-        case .withdraw:
-            showInfo = onlineAccountState.isLinkingInProgress
-        default:
-            showInfo = false
-        }
-
-        additionalInfo.isHidden = !showInfo
-        showInfoConstraint.isActive = showInfo
-        collapseInfoConstraint.isActive = !showInfo
-        infoBottomAnchorConstraint.isActive = showInfo
-
-        if showInfo {
-            additionalInfo.backgroundColor = item.infoBackgroundColor
-            additionalInfoLabel.text = item.info(crowdNodeBalance, onlineAccountState)
-            additionalInfoLabel.textColor = item.infoTextColor
-
-            if !item.infoActionButton(for: onlineAccountState).isEmpty {
-                additionalInfoLabel.textAlignment = .left
-                additionalInfoIcon.isHidden = false
-                verifyButton.isHidden = false
-            } else {
-                additionalInfoLabel.textAlignment = .center
-                additionalInfoIcon.isHidden = true
-                verifyButton.isHidden = true
-            }
-        }
-    }
-}
-
-// MARK: - CrowdNodePortalController + UITableViewDelegate, UITableViewDataSource
+// MARK: UITableViewDelegate, UITableViewDataSource
 
 extension CrowdNodePortalController : UITableViewDelegate, UITableViewDataSource {
     func numberOfSections(in tableView: UITableView) -> Int {
@@ -323,7 +255,7 @@ extension CrowdNodePortalController : UITableViewDelegate, UITableViewDataSource
         case .withdraw:
             let account = DWEnvironment.sharedInstance().currentAccount
             let allAvailableFunds = account.maxOutputAmount
-            
+
             if allAvailableFunds >= CrowdNode.minimumLeftoverBalance {
                 navigationController?.pushViewController(CrowdNodeTransferController.controller(mode: TransferDirection.withdraw), animated: true)
             } else {
@@ -337,21 +269,22 @@ extension CrowdNodePortalController : UITableViewDelegate, UITableViewDataSource
             UIApplication.shared.open(URL(string: CrowdNode.supportUrl)!)
         }
     }
-    
+
     private func showMinimumBalanceError() {
         let vc = BasicInfoController()
         vc.icon = "image.crowdnode.error"
         vc.headerText = NSLocalizedString("You should have a positive balance on Dash Wallet", comment: "CrowdNode")
-        vc.descriptionText = String.localizedStringWithFormat(NSLocalizedString("Deposit at least %@ Dash on your Dash Wallet to complete a withdrawal", comment: "CrowdNode"), CrowdNode.minimumLeftoverBalance.formattedDashAmountWithoutCurrencySymbol)
-        
-        if (DWEnvironment.sharedInstance().currentChain.isMainnet()) {
+        vc.descriptionText = String.localizedStringWithFormat(NSLocalizedString("Deposit at least %@ Dash on your Dash Wallet to complete a withdrawal", comment: "CrowdNode"),
+                                                              CrowdNode.minimumLeftoverBalance.formattedDashAmountWithoutCurrencySymbol)
+
+        if DWEnvironment.sharedInstance().currentChain.isMainnet() {
             vc.actionButtonText = NSLocalizedString("Buy Dash", comment: "CrowdNode")
         } else {
             vc.actionButtonText = ""
         }
-        
+
         let nvc = BaseNavigationController(rootViewController: vc)
-        
+
         vc.mainAction = {
             Task {
                 if await self.viewModel.authenticate() {
@@ -360,12 +293,12 @@ extension CrowdNodePortalController : UITableViewDelegate, UITableViewDataSource
                 }
             }
         }
-        
+
         present(nvc, animated: true)
     }
 }
 
-// MARK: - CrowdNodePortalController + BalanceViewDataSource
+// MARK: BalanceViewDataSource
 
 extension CrowdNodePortalController: BalanceViewDataSource {
     var mainAmountString: String {
