+++ DashWallet/Sources/UI/CrowdNode/Portal/StakingInfoDialogController.swift
@@ -46,6 +46,8 @@ final class StakingInfoDialogController: UIViewController {
     }
 
     private func configureHierarchy() {
+        view.backgroundColor = .dw_secondaryBackground()
+
         addressLabel.text = viewModel.accountAddress
         let minimumDeposit = DSPriceManager.sharedInstance().string(forDashAmount: Int64(CrowdNode.minimumDeposit))!
         minimumDepositLabel.text = String.localizedStringWithFormat(NSLocalizedString("You only need %@ to join the pool.", comment: "CrowdNode"), minimumDeposit)
